requirements.txt
requests>=2.32.0
paho-mqtt>=1.6.0
pycryptodome>=3.20.0
pyyaml>=6.0.1
qrcode>=7.4.2
pillow>=10.0.0
tzlocal>=5.2
